Fix calculation of inner_d and inner_l
Add parentheseis around the nested question mark operators to fix
calculation of the inner_d and inner_l variables. Without the
parentheses, extra_d and extra_l were ignored.

Also move the declaration of extra_d and extra_l and the derived
values further down, so as to hide them from the customizer.